+++ littletree/node.py
@@ -101,10 +101,10 @@ class Node(BaseNode[TIdentifier], Generic[TIdentifier, TData]):
             while insert_depth >= item.depth:
                 insert_depth -= 1
                 diff_node = diff_node.parent
-            diff_node[node.identifier] = diff_node = Node()
+            diff_node = diff_node.path.create([node.identifier])
             diff_node.data['other'] = node.data
             insert_depth += 1
-        diff_tree = diff_node.root
+        diff_tree = diff_node.get_root()
         if not keep_equal:
             to_detach = None
             for node in diff_tree.iter_tree(order='post'):